Issue #347: Restructure directories
