+++ fitbenchmarking/controllers/mantid_controller.py
@@ -7,7 +7,7 @@ from mantid.api import *
 from mantid.fitfunctions import *
 import numpy as np
 
-from fitbenchmarking.fitting.controllers.base_controller import Controller
+from fitbenchmarking.controllers.base_controller import Controller
 
 
 class MantidController(Controller):
