+++ fitbenchmarking/controllers/sasview_controller.py
@@ -7,7 +7,7 @@ from bumps.names import Curve, FitProblem
 
 import numpy as np
 
-from fitbenchmarking.fitting.controllers.base_controller import Controller
+from fitbenchmarking.controllers.base_controller import Controller
 
 
 class SasviewController(Controller):
