+++ fitbenchmarking/core/fitting_benchmarking.py
@@ -10,7 +10,7 @@ from fitbenchmarking.utils.logging_setup import logger
 
 from fitbenchmarking.parsing.parser_factory import parse_problem_file
 from fitbenchmarking.utils import create_dirs, misc
-from fitbenchmarking.fitbenchmark_one_problem import fitbm_one_prob
+from fitbenchmarking.core.fitbenchmark_one_problem import fitbm_one_prob
 
 
 def fitbenchmark_group(group_name, options, data_dir):
