+++ fitbenchmarking/core/results_output.py
