+++ fitbenchmarking/plotting/plots.py
